[bugfix][ChromX][cxachartwidget] fixed current Test runTime data format
[what] 1、修复当前运行时间数据格式

[why] 原来quint32整形，以毫秒为单位。需要改为以秒为单位的浮点型

[how] 变更UI和原始数据转换方法
      变更数据类型和显示方法
      还有csv导入导出

+++ src/ChromXApp/TransUiValueConvertor/cxatestdataconvertor.h
@@ -6,7 +6,7 @@
 typedef struct tagUiTestData {
     double TDCurTemperature;
     double TICurTemperature;
-    quint32 curTestRunTime;
+    double curTestRunTime;
     double COLUMNTemperature;
     double MicroPIDValue;
